+++ ui/components/popover-header/popover.ts
@@ -1,0 +1,1 @@
+export * from './popover-header.component'
